refactor: fix import paths and rmeove redundant gcsFilesystemOptions
